% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_O2}
\alias{a_O2}
\title{Oxygen (O2) absorption spectrum}
\format{
\subsection{\code{a_O2}}{

A data frame with XXX row and two columns:
\describe{
\item{wavelength}{radiation wavelength (nm)}
\item{a_O2}{absorption coefficient of oxygen (cm-1)}
}
}
}
\source{
WASI4 validate with Soham
}
\usage{
a_O2
}
\description{
Oxygen (O2) absorption spectrum
}
\keyword{datasets}
