% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saber_class.R
\name{blacklist}
\alias{blacklist}
\title{Extract motif/variant blacklist from a \code{Saber} object}
\usage{
blacklist(x, ...)
}
\arguments{
\item{x}{A \code{Saber} object.}
}
\value{
For a \code{Saber} object, a character vector of variant IDs
corresponding to stereotyped recurrent and/or "no variant" motifs that
should be treated as background.
}
\description{
Generic accessor for retrieving the blacklist of uninformative barcodes
(variants) from a \code{Saber} object. For \code{Saber}, this returns the
\code{@motif_blacklist} slot.
}
\seealso{
\code{\link{Saber}}, \code{\link{make.Saber}()}
}
\keyword{internal}
