% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanhellemont_2018.R
\name{calc_rho_sky}
\alias{calc_rho_sky}
\title{calculate the sky reflectance (sky glint, rho_sky) at the air-water interface}
\usage{
calc_rho_sky(
  rrs_subsurface,
  wavelength,
  view_above = 15,
  sun_above = 30,
  q_surf = pi,
  g_dd = 0.05,
  g_dsr = 0,
  g_dsa = 0,
  f_dd = 1,
  f_ds = 1,
  alpha = 1.317,
  P = 1013.25,
  RH = 0.6,
  Hoz = 0.3,
  WV = 2.5
)
}
\description{
calculate the sky reflectance (sky glint, rho_sky) at the air-water interface
}
\references{
Gege, P. (2012)
‘Analytic model for the direct and diffuse components of downwelling
spectral irradiance in water’, Applied Optics, 51(9), p. 1407.
Available at: https://doi.org/10.1364/AO.51.001407.

Vanhellemont, Q. and Ruddick, K. (2018)
‘Atmospheric correction of metre-scale optical satellite data for
inland and coastal water applications’, Remote Sensing of Environment,
216, pp. 586–597.
Available at: https://doi.org/10.1016/j.rse.2018.07.015.
}
\author{
Soham Mukherjee
}
