% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_r_rs_b.R
\name{compute_r_rs_b_lmm}
\alias{compute_r_rs_b_lmm}
\title{Compute interpolated Rrs_b mixture}
\usage{
compute_r_rs_b_lmm(fractions, wavelength)
}
\arguments{
\item{fractions}{named numeric vector (e.g., c(sand = 0.6, seagrass = 0.4))}

\item{wavelength}{numeric vector}
}
\description{
Compute interpolated Rrs_b mixture
}
