% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_fct.R
\name{create_prior_distribution}
\alias{create_prior_distribution}
\title{Create prior distribution}
\usage{
create_prior_distribution(
  use_prior = c("ioccg", "nomad", "wiseman"),
  plot_diag = F,
  distribution_fit = c("weibull", "lognormal", "normal"),
  truncate_chl = c(0.5, 30),
  truncate_adg = c(0.1, 5),
  truncate_bbp = c(0.002, 0.01),
  sample_count = 100
)
}
\arguments{
\item{use_prior}{choose dataset from wich to construct prior}
}
\value{
input data prior distribution to be used by ?
}
\description{
Create prior distribution
}
\author{
Soham Mukherjee
}
