% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{.error_function_registry}
\alias{.error_function_registry}
\title{Error Function Registry}
\format{
An object of class \code{environment} of length 3.
}
\usage{
.error_function_registry
}
\description{
List of error functions available to SABER
}
\keyword{datasets}
