% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{.objective_function_registry}
\alias{.objective_function_registry}
\title{Error Function Registry}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.objective_function_registry
}
\description{
List of error functions available to SABER
}
\keyword{datasets}
