% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_objective.R
\name{factory_objective}
\alias{factory_objective}
\title{Make objective function}
\usage{
factory_objective(model, error, rrs_observed, par_fixed = NULL)
}
\arguments{
\item{model}{Name of the forward model}

\item{error}{Name of the error function}

\item{rrs_observed}{Observed Rrs data}

\item{par_fixed}{Named list or data frame of fixed parameters}
}
\value{
A function returning the model error/likelihood to be used in an optimization process
}
\description{
Make objective function
}
