% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_minimization.R
\name{fct_compose_minimization}
\alias{fct_compose_minimization}
\title{fct_compose_minimization
a function to dynamically create the minimization function
passed to inversion}
\usage{
fct_compose_minimization(forward_model, error_fct, rrs_observed)
}
\arguments{
\item{forward_model}{c("am03", "lee98")}

\item{error_fct}{c("log-ll", "lee99")}
}
\description{
fct_compose_minimization
a function to dynamically create the minimization function
passed to inversion
}
\author{
Raphael Mabit
}
