% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gege_2012.R
\name{gege_2012}
\alias{gege_2012}
\title{Compute Rayleigh scattering at the air-sea interface (sky glint)}
\usage{
gege_2012(
  wavelength,
  theta_s,
  g_dd = 0.02,
  g_dsr = 1/pi,
  g_dsa = 1/pi,
  f_dd = 1,
  f_ds = 1,
  AM = 1,
  alpha = 1.317,
  P = 1013.25,
  RH = 0.6,
  Hoz = 0.3,
  WV = 2.5
)
}
\arguments{
\item{alpha}{angstrom exponent}

\item{RH}{relative humidity}

\item{Hoz}{scale height for ozone ? \link{cm} or \link{cm-atm}}

\item{WV}{scale height water vapor ? \link{cm} or \link{cm-atm}}

\item{p}{atmospheric pressure \link{mb}}
}
\description{
Compute Rayleigh scattering at the air-sea interface (sky glint)
}
\references{
Gordon, H.R., Brown, J.W. and Evans, R.H. (1988)
‘Exact Rayleigh scattering calculations for use with the Nimbus-7
Coastal Zone Color Scanner’, Applied Optics, 27(5), p. 862.
Available at: https://doi.org/10.1364/AO.27.000862.

Gege, P. (2012)
‘Analytic model for the direct and diffuse components of downwelling
spectral irradiance in water’, Applied Optics, 51(9), p. 1407.
Available at: https://doi.org/10.1364/AO.51.001407.
}
\author{
Soham Mukerjee, Raphael Mabit
}
