% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gestalt_typing}
\alias{gestalt_typing}
\title{Type Gestalt Barcoding DAta}
\usage{
gestalt_typing(
  output_folder,
  config,
  trimmomatic_path = "/opt/Trimmomatic-0.39/trimmomatic-0.39.jar"
)
}
\arguments{
\item{output_folder}{Folder for output}

\item{config}{Config file for setting up the analysis.  See extdata/gestalt_config.csv}

\item{trimmomatic_path}{PARAM_DESCRIPTION, Default: '/opt/Trimmomatic-0.39/trimmomatic-0.39.jar'}
}
\value{
Does not return anything
}
\description{
Generates typing results in the specified directory.  A time stamp is appended to the directory path provided.
}
