% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccesstoken.R
\name{getAccesstoken}
\alias{getAccesstoken}
\title{Get your access token}
\usage{
getAccesstoken(api_key, secret_key)
}
\arguments{
\item{api_key}{The API Key}

\item{secret_key}{The Secret Key}
}
\value{
access token.
}
\description{
Get your access token
}
\examples{
\dontrun{
access_token = getAccesstoken(api_key = "xxxxxxxx", secret_key = "xxxxxxxx")
}

}
\references{
https://cloud.baidu.com/doc/NLP/s/Qk3pmn3s7
}
