% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAddress.R
\name{getAddress}
\alias{getAddress}
\title{Standardization of address}
\usage{
getAddress(data, address, confidence = 100, retry = 3)
}
\arguments{
\item{data}{Data set}

\item{address}{The column name of address}

\item{confidence}{This parameter is used to trigger the supplementary resolution strategy, and perform supplementary resolution on the results with confidence below the configuration value to improve the accuracy of the results. confidence = 100 by default.}

\item{retry}{The times of retry when return empty results. retry = 3 by default.}
}
\value{
data.table
}
\description{
Standardization of address using Baidu natural language processing api easily in R
}
\examples{
\dontrun{
results <- getAddress(data = df, address = "addr")
}

}
\references{
https://cloud.baidu.com/doc/NLP/s/Qk3pmn3s7
}
