% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_inverse.R
\name{inverse_mcmc}
\alias{inverse_mcmc}
\title{Markov Chain Monte Carlo radiative transfer inversion}
\usage{
inverse_mcmc(density, sampler, likelihood)
}
\description{
Markov Chain Monte Carlo radiative transfer inversion
}
\author{
Soham Mukherjee
}
