% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversion_gradient.R
\name{inversion_gradient}
\alias{inversion_gradient}
\title{Perform inversion based on various optimization methods.
Allow for unconstrained inversion of chl \link{mg/m^-3}, ag_440 \link{m^-1},
bbp_550 \link{m^-1}, and in optically shallow waters h_w \link{m},
and bottom reflectance fraction. Optionally, can perform constrained
inversion by providing parameters in \code{fixed_par} instead of \code{par_init}.}
\usage{
inversion_gradient(
  rrs,
  forward_model,
  objective_fct,
  optim_mtd,
  par_init,
  par_fixed = NULL,
  lower_b = NULL,
  upper_b = NULL,
  verbose = F
)
}
\arguments{
\item{forward_model}{c("am03", "lee98")}

\item{objective_fct}{c("log-ll", "SSR", "lee99") SSR is not implemented in the code}

\item{optim_mtd}{c("nelder-mead", "BFGS", "CG", "L-BFGS-B", "sann", "brent","levenberg-marqardt", "auglag")}

\item{par_init}{a tibble with the initial parameter to be inverted.
For best results, can be estimated with \code{pre_fit_inversion}.
\describe{
\item{chl}{chlorophyl-a concentration in \link{mg/m^3}}
\item{ag_440}{CDOM absorption \link{m^-1} at 440 nm}
\item{bbp_550}{particulate backscattering \link{m^-1} at 550 nm}
\item{h_w}{watercolumn height above the bottom \link{m}}
\item{rb_*}{fraction of end-member bottom reflectance class}
\item{sd}{Optional, standard deviaton of the population for \code{objective_fct = "log-ll"}}
}}

\item{lower_b}{Optional, lower boundary for \code{optim_mtd = "L-BFGS-B"}.
If not provided will be calculated in \code{parse_inverse_parameter}}

\item{upper_b}{Optional, same as \code{lower_b}.}

\item{verbose}{guess}

\item{fixed_par}{Optional, a tibble with the same columns as par_init.
Parameter defined here will be considered known, hence not retrieved during
optimization. If defined here they must not be defined in \code{par_init}.}
}
\description{
Perform inversion based on various optimization methods.
Allow for unconstrained inversion of chl \link{mg/m^-3}, ag_440 \link{m^-1},
bbp_550 \link{m^-1}, and in optically shallow waters h_w \link{m},
and bottom reflectance fraction. Optionally, can perform constrained
inversion by providing parameters in \code{fixed_par} instead of \code{par_init}.
}
\author{
Soham Mukherjee, Raphael Mabit
}
