% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversion_unconstrained.R
\name{inversion_unconstrained}
\alias{inversion_unconstrained}
\title{inversion_unconstrained
Unconstrained full inversion. Two options for forward semi analytical model
and two option for the objective function.}
\usage{
inversion_unconstrained(
  rrs,
  init_par,
  forward_model,
  error_fct,
  optim_mtd,
  lower_b = NULL,
  upper_b = NULL,
  verbose = F
)
}
\arguments{
\item{forward_model}{c("am03", "lee98")}

\item{error_fct}{c("log-ll", "SSR", "lee99") SSR is not implemented in the code}

\item{optim_mtd}{c("nelder-mead", "BFGS", "CG", "L-BFGS-B", "sann", "brent","levenberg-marqardt", "auglag")}
}
\description{
inversion_unconstrained
Unconstrained full inversion. Two options for forward semi analytical model
and two option for the objective function.
}
\author{
Soham Mukherjee
}
