% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iop_from_oac.R
\name{iop_from_oac}
\alias{iop_from_oac}
\title{iop_from_oac
compute Inherent Optical Proprieties from Optically Active constituents.}
\usage{
iop_from_oac(
  wavelength,
  oac,
  rrs = NULL,
  optically_shallow = NULL,
  verbose = F
)
}
\arguments{
\item{oac}{a tibble with column {chl \link{mg/m^-3}, ag_440, bb_550}, optionally,
provide manual ag and bb slope columns {ag_s_g, ag_s_d, bb_gamma}}

\item{rrs}{optionnal, a tibble with columns {wavelength, rrs} used for
parametric retrieval of a_g and bb slope}
}
\description{
iop_from_oac
compute Inherent Optical Proprieties from Optically Active constituents.
}
