% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iop_from_oac_spm}
\alias{iop_from_oac_spm}
\title{Compute IOPs from OACs}
\usage{
iop_from_oac_spm(
  wavelength,
  a_w,
  a0,
  a1,
  bb_w,
  chl,
  a_g_440,
  spm,
  a_nap_star,
  bb_p_star,
  a_g_s,
  a_nap_s,
  bb_p_gamma
)
}
\description{
Bio-optical model to compute Inherent Optical Proprieties from
chl, g+nap(440), bb_p_550
}
