% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_forward.R
\name{lee98_forward}
\alias{lee98_forward}
\title{Lee 1998 forward model
Compute Rrs given IOP}
\usage{
lee98_forward(
  wavelength,
  iop,
  view,
  sun,
  optically_shallow,
  h_w = NULL,
  rrs_bottom = NULL,
  verbose = F
)
}
\arguments{
\item{wavelength}{Requested wavelength for the simulation}

\item{iop}{a tibble with column {wavelength, a, bb}, non-water component of absorption and backscattering.}

\item{optically_shallow}{bolean, is the water optically shallow ? If TRUE, provide \code{h_w}, \code{rb_fraction}, \code{rb}.}

\item{h_w}{water column height \link{m}.}

\item{rrs_botom}{remote sensing bottom reflectance. see function \code{rrs_bottom_lmm}}
}
\description{
Lee 1998 forward model
Compute Rrs given IOP
}
\references{
Lee, z. et al. (1998) ‘Hyperspectral remote sensing for shallow waters. I. A semianalytical model’, Applied Optics, 37(27), pp. 6329–6338. Available at: https://doi.org/10.1364/AO.37.006329.
}
\author{
Soham Mukherjee
}
