% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_likelihood.R
\name{make_likelihood}
\alias{make_likelihood}
\title{Make likelihood/minimization function}
\usage{
make_likelihood(model, error, rrs_observed, par_fixed = NULL)
}
\arguments{
\item{model}{Name of the forward model}

\item{error}{Name of the error function}

\item{rrs_observed}{Observed Rrs data}

\item{par_fixed}{Named list or data frame of fixed parameters}
}
\value{
A function returning the model error/likelihood
}
\description{
Make likelihood/minimization function
}
