% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_stan_data.R
\name{make_stan_data}
\alias{make_stan_data}
\title{Make stan data}
\usage{
make_stan_data(
  stan_data_base,
  stan_data_obs,
  h_w_mu_from_obs = F,
  h_w_sd_from_obs = F
)
}
\description{
Make stan data
}
