% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_stan_data.R
\name{make_stan_data_base}
\alias{make_stan_data_base}
\title{Make base data needed by STAN model}
\usage{
make_stan_data_base(
  wavelength,
  water_type,
  theta_view,
  shallow,
  bottom_class_names,
  ...
)
}
\arguments{
\item{...}{additional fixed data expected by the model}
}
\description{
Make base data needed by STAN model
}
