% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_stan_data.R
\name{make_stan_data_base_gmm}
\alias{make_stan_data_base_gmm}
\title{Make base data needed by STAN model (low-rank PCA mixture bottom prior in ETA/logit space)}
\usage{
make_stan_data_base_gmm(
  wavelength,
  water_type,
  theta_view,
  shallow,
  bottom_class_names,
  rb_obs_path = "/home/raphael/R/SABER/inst/extdata/r_b_gamache_obs.csv",
  q = 5,
  use_empirical_pi = FALSE,
  ...
)
}
\description{
Make base data needed by STAN model (low-rank PCA mixture bottom prior in ETA/logit space)
}
