% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_stan_data.R
\name{make_stan_data_base_splines}
\alias{make_stan_data_base_splines}
\title{Make base data needed by STAN model (spline mixture bottom prior)}
\usage{
make_stan_data_base_splines(
  wavelength,
  water_type,
  theta_view,
  shallow,
  bottom_class_names,
  rb_obs_path = "/home/raphael/R/SABER/inst/extdata/r_b_gamache_obs.csv",
  m_spline = 15,
  spline_degree = 3,
  use_empirical_pi = TRUE,
  ...
)
}
\description{
Make base data needed by STAN model (spline mixture bottom prior)
}
