% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parser.R
\name{memoised_a0_a1_phyto}
\alias{memoised_a0_a1_phyto}
\title{function to memoise internal package data at requested wavelength
speed up computation by avoiding repetitive interpolation}
\usage{
memoised_a0_a1_phyto(wavelength)
}
\description{
function to memoise internal package data at requested wavelength
speed up computation by avoiding repetitive interpolation
}
