% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parser.R
\name{parse_inverse_parameter}
\alias{parse_inverse_parameter}
\title{parse_inverse_parameter}
\usage{
parse_inverse_parameter(
  par_df,
  optim_mtd,
  lower_b = NULL,
  upper_b = NULL,
  verbose = F
)
}
\description{
a function to parse and prepare the init parameter to be passed to optimization function
}
