% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parser.R
\name{parse_iop}
\alias{parse_iop}
\title{parse_iop}
\usage{
parse_iop(a, bb, wavelength, verbose = F)
}
\arguments{
\item{a}{tibble with columns {a, wavelength}}

\item{bb}{tibble with columns {bb, wavelength}}

\item{wavelength_out}{a vector with the requested interpolation wavelength}

\item{plot}{bollean, produce input and interpolated IOPs plot}
}
\value{
A tibble with columns {wavelength, a, bb}
}
\description{
take tibbles for non water absorption and backscattering
}
\author{
Raphael Mabit
}
