% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_saber.R
\name{plot.saber}
\alias{plot.saber}
\title{Plot function for \code{saber} objects

TODO}
\usage{
\method{plot}{saber}(x, h = 3, v = 2, shape = "circle",
  edge.width = 1, testNiveaus = c(1e-10, 1e-05, 0.01, 0.5), ...)
}
\arguments{
\item{x}{[\\code{saber.result}] \cr
TODO}

\item{h}{[\code{integer(1)}] \cr
TODO}

\item{v}{[\code{integer(1)}] \cr
TODO}

\item{shape}{[\code{character(1)}] \cr
TODO}

\item{edge.width}{[\code{integer(1)}] \cr
TODO}

\item{testNiveaus}{[\code{double()}] \cr
TODO}

\item{...}{[any] \cr
Not used.}
}
\value{
Nothing, as a side effect, a plot is generated.
}
\description{
Plot function for \code{saber} objects

TODO
}
