% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_saber.R
\name{print.saber}
\alias{print.saber}
\title{Print function for SABeR objects}
\usage{
\method{print}{saber}(x, ...)
}
\arguments{
\item{x}{[\code{saber}] \cr
SABeR object}

\item{...}{\cr
Not used}
}
\description{
Print function for SABeR objects
}
