% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_r_b.R
\name{r_b_lmm}
\alias{r_b_lmm}
\title{r_b_lmm
Compute bottom remote sensing reflectance based on a Linear Mixture Model
of end member spectra.}
\usage{
r_b_lmm(rrs_end_member, end_member_fraction, wavelength, verbose = FALSE)
}
\arguments{
\item{rrs_end_member}{A tibble with columns {wavelength, class, r_b} for end member spectra}

\item{end_member_fraction}{A tibble with columns {class, fraction} that sum to 1}

\item{wavelength}{Target wavelengths for interpolation}

\item{verbose}{If TRUE, plots original and interpolated spectra}
}
\value{
A tibble with columns {wavelength, r_b}
}
\description{
r_b_lmm
Compute bottom remote sensing reflectance based on a Linear Mixture Model
of end member spectra.
}
