% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_rs_b_gamache}
\alias{r_rs_b_gamache}
\title{Bottom reflectance data collected during the Algae-WISE mission}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1500 rows and 4 columns.
}
\source{
Unpublished Dataset
}
\usage{
r_rs_b_gamache
}
\description{
This data was collected in the southwestern shores of Anticosti Island
in 2022 and 2023. It represent different end-members spectrum and composite
spectrum of bottom substrate and algal species.
}
\details{
@format ## \code{rb_algae_wise}
A data frame with XXX rows and XXX columns:
\describe{
\item{wavelength}{wavelength of bottom reflectance in nanometer}
\item{class}{class of bottom reflectance end-member spectrum}
\item{rb}{bottom reflectance}
}
}
\keyword{datasets}
