% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refraction_law.R
\name{refraction_law}
\alias{refraction_law}
\title{refraction_law}
\usage{
refraction_law(incidence_angle, n1, n2)
}
\arguments{
\item{incidence_angle}{angle of incidence of electromagnetic radiation (rad)}

\item{n_1}{refractive index of media of incidence (e.g. n_air = 1)}

\item{n_2}{refractive index of media of transmittance (e.g. n_water = 1.34)}
}
\description{
refraction_law
}
\references{
https://en.wikipedia.org/wiki/Snell\%27s_law
}
\author{
Soham Murkerjee
}
