% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_error_function}
\alias{register_error_function}
\title{Register a new error function}
\usage{
register_error_function(name, fn)
}
\arguments{
\item{name}{Name of the error function}

\item{fn}{A function that returns a single numeric error value}
}
\description{
Register a new error function
}
