% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_forward_model}
\alias{register_forward_model}
\title{Register a new forward model}
\usage{
register_forward_model(name, fn)
}
\arguments{
\item{name}{Name of the model}

\item{fn}{A function taking model parameters and returning modeled Rrs}
}
\description{
Register a new forward model
}
