% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_input_preparer}
\alias{register_input_preparer}
\title{Register a new input preparation}
\usage{
register_input_preparer(name, fn)
}
\arguments{
\item{name}{Name of the input preparation (must end with *input_\*modelName\**)}

\item{fn}{A function taking model parameters and returning model inputs}
}
\description{
Register a new input preparation
}
