% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_objective_function}
\alias{register_objective_function}
\title{Register a new objective function}
\usage{
register_objective_function(name, fn)
}
\arguments{
\item{name}{Name of the error function}

\item{fn}{A function that returns a single numeric error value}
}
\description{
Register a new objective function
}
