% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectance_bottom.R
\name{rrs_bottom_lmm}
\alias{rrs_bottom_lmm}
\title{rrs_bottom_lmm
Compute bottom remote sensing reflectance based on Linear Mixture Model
of end member spectra.}
\usage{
rrs_bottom_lmm(rrs_end_member, end_member_fraction, wavelength, verbose = F)
}
\value{
a tibble with columns {wavelength, rrs_bottom}
@export
}
\description{
@param rrs_end_member a tibble with columns {wavelength, class, rrs_b} end members spectrum
@param rb_fraction a tibble with columns {class, fraction}, fraction must sum to 1
@param wavelength requested interpolation wavelength
}
