% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrs_subsurface_to_surface.R
\name{rrs_subsurface_to_surface}
\alias{rrs_subsurface_to_surface}
\title{Transmit Rrs(0-) to Rrs(0+)}
\usage{
rrs_subsurface_to_surface(
  rrs_subsurface,
  wavelength,
  view_above = 15,
  sun_above = 30,
  q_surf = pi
)
}
\description{
Transmit Rrs(0-) to Rrs(0+)
}
\references{
Gege, P. (2012)
‘Analytic model for the direct and diffuse components of downwelling
spectral irradiance in water’, Applied Optics, 51(9), p. 1407.
Available at: https://doi.org/10.1364/AO.51.001407.
}
\author{
Soham Mukherjee
}
