% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{saber-package}
\alias{saber}
\alias{saber-package}
\title{\packageTitle{saber}}
\description{\packageDescription{saber}}
\section{Package Content}{\packageIndices{saber}}
\author{\packageAuthor{saber}}
\section{Maintainer}{\packageMaintainer{saber}}
\keyword{internal}
