% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saberIt.R
\name{saberIt}
\alias{saberIt}
\title{Statistical Analysis of Benchmark Results}
\usage{
saberIt(data, perfName, expParName, algoName, replName,
  method = "kmeans", min.in.cluster = n.exps * 0.1, max.nc = n.exps *
  0.5, ...)
}
\arguments{
\item{data}{[\code{data.frame}] \cr
Data.frame containing the data to be analyzed}

\item{perfName}{[\code{character(1)}] \cr
Name of the column of data containing the performance values}

\item{expParName}{[\code{character()}] \cr
Names of the columns of data containing the experiment parameters}

\item{algoName}{[\code{character(1)}] \cr
Name of the column of data containing the used algorithm}

\item{replName}{[\code{character(1)}] \cr
Name of the column of data containing the replication number}

\item{method}{[\code{character(1)}] \cr
method used in \code{NbClust}}

\item{min.in.cluster}{[\code{numeric(1)}]\cr
Each found cluster shall contain at least this number of observations.}

\item{max.nc}{[\code{numeric(1)}]\cr
Maximum number of clusters.}

\item{...}{\cr
Additional parameters for \code{NbClust}}
}
\value{
A \code{saber}, object.
}
\description{
TODOTODOTODOTODOTODOTODO
}
