% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sample_dataset.R
\name{sample_nomad}
\alias{sample_nomad}
\title{sample_nomad}
\usage{
sample_nomad(
  truncate_chl = NULL,
  truncate_ag = NULL,
  truncate_ad = NULL,
  truncate_bbp = NULL,
  sample_count = 100
)
}
\value{
input data prior distribution to be used by ?
}
\description{
sample the NOMAD dataset
}
\author{
Soham Mukherjee
}
