% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_objective_inverse.R
\name{solve_objective_inverse_shallow}
\alias{solve_objective_inverse_shallow}
\title{saber_inverse
Implement the SABER inversion. BGC stands for biogeochemstry}
\usage{
solve_objective_inverse_shallow(
  constrain_shallow_iop,
  abs_path,
  bb_path,
  constrained_bgc,
  constrain_bgc_param = c("chl", "adg", "bbp"),
  constrain_bgc_value,
  unconstrained,
  auto_spectral_slope,
  manual_spectral_slope,
  manual_spectral_slope_vals = c(s_g = 0.015, s_d = 0.0116, gamma = 0.5),
  initial,
  objective_fct = c("log-LL", "SSR", "obj_L98"),
  sa_model = "am03",
  method_opt = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent",
    "levenberg-marqardt", "auglag"),
  obsdata,
  wavelength,
  lower_b,
  upper_b
)
}
\arguments{
\item{objective_fct}{c("log-LL", "SSR", "obj_L98")}

\item{sa_model}{c("am03", "lee98")}

\item{method_opt}{c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent","levenberg-marqardt", "auglag")}
}
\description{
saber_inverse
Implement the SABER inversion. BGC stands for biogeochemstry
}
