% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCluster.R
\name{summarizeCluster}
\alias{summarizeCluster}
\title{summarizeCluster}
\usage{
summarizeCluster(saber.result)
}
\arguments{
\item{saber.result}{[\code{saber}] \cr
Result of saberIt}
}
\value{
Summary printed on the console as a side effect, a ggplot object is returned
}
\description{
Summary of the clusters found in the saber analysis
}
