% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_input_rrs}
\alias{test_input_rrs}
\title{Remote sensing reflectance}
\format{
\subsection{\code{test_input_rrs}}{

A data frame with 146 row and two columns:
\describe{
\item{wavelength}{radiation wavelength (nm)}
\item{Rrs}{remote sensing reflectance above surface (0+) (sr-1)}
}
}
}
\source{
MAN-F21 2019-08-18 22:50:00  49.0 -68.3 PSR
}
\usage{
test_input_rrs
}
\description{
Remote sensing reflectance
}
\keyword{datasets}
