% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{.input_preparer_registry}
\alias{.input_preparer_registry}
\title{Forward model input preparation Registry}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.input_preparer_registry
}
\description{
List of forward modelsinput preparation available to SABER
}
\keyword{datasets}
