% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward.R
\name{forward_am03}
\alias{forward_am03}
\title{Albert and Mobley 2003 forward model}
\usage{
forward_am03(
  wavelength,
  iop,
  water_type = 2,
  theta_view,
  theta_sun,
  optically_shallow = F,
  h_w = NULL,
  r_b = NULL,
  verbose = F
)
}
\description{
Fuction repackaged from \code{SABER_forward_fast.R} in codebase. What's the
difference with \code{saber_forward_parametric_conc_wise.R} ?
}
\references{
Albert, A. and Mobley, C.D. (2003) ‘An analytical model for subsurface irradiance and remote sensing reflectance in deep and shallow case-2 waters’, Optics Express, 11(22), pp. 2873–2890. Available at: https://doi.org/10.1364/OE.11.002873.
}
