% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversion_unconstrained.R
\name{inversion_unconstrained}
\alias{inversion_unconstrained}
\title{inversion_unconstrained
Unconstrained full inversion. Two options for forward semi analytical model
and two option for the objective function.}
\usage{
inversion_unconstrained(wavelength, sa_model, objective_fct)
}
\arguments{
\item{sa_model}{c("am03", "lee98")}

\item{objective_fct}{c("log-LL", "SSR", "lee99") SSR is not implemented in the code}

\item{method_opt}{c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent","levenberg-marqardt", "auglag")}
}
\description{
inversion_unconstrained
Unconstrained full inversion. Two options for forward semi analytical model
and two option for the objective function.
}
\author{
Soham Mukherjee
}
