% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iop_from_oac}
\alias{iop_from_oac}
\title{Compute IOPs from OACs}
\usage{
iop_from_oac(
  wavelength,
  a_w,
  a0,
  a1,
  bb_w,
  chl,
  a_gnap_440,
  bb_p_550,
  a_gnap_s,
  bb_p_gamma
)
}
\description{
Bio-optical model to compute Inherent Optical Proprieties from Optically Active Constituents
}
