% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee_forward.R
\name{lee98_forward}
\alias{lee98_forward}
\title{Lee 1998 forward model
Compute Rrs given initial parameters
Question For Soham:
Where does the parametric formula to retrieve spectral slope of CDOM + NAP comes from, QAA ?
Where does the a_w values comes from ?
Why is pure water backscattering dependent on water type ?}
\usage{
lee98_forward(
  iop = NULL,
  oac = NULL,
  water_type = 2,
  optically_shallow = T,
  h_w = 2,
  rb_fraction,
  rb = read_csv(fs::path_package("SABER", "extdata", "rb_endmembers_algae-wise.csv")),
  view = 40,
  sun = 20,
  wavelength = seq(400, 800, 10),
  rrs = NULL,
  verbose = T,
  plot = T
)
}
\arguments{
\item{iop}{a tibble with column {wavelength, a, bb}}

\item{oac}{a tibble with column {chl \link{mg/m^-3}, ag_440, bb_550}, optionally,
to provide manual ag and bb slope columns {ag_s_g, ag_s_d, bb_gamma}}

\item{rb}{optional, a tibble with columns {wavelength, class\code{X}} end member
reflectance spectra.}

\item{rrs}{optionnal, a tibble with columns {wavelength, rrs} used for
parametric retrieval of a_g and bb slope}
}
\description{
Lee 1998 forward model
Compute Rrs given initial parameters
Question For Soham:
Where does the parametric formula to retrieve spectral slope of CDOM + NAP comes from, QAA ?
Where does the a_w values comes from ?
Why is pure water backscattering dependent on water type ?
}
\references{
Lee, z. et al. (1998) ‘Hyperspectral remote sensing for shallow waters. I. A semianalytical model’, Applied Optics, 37(27), pp. 6329–6338. Available at: https://doi.org/10.1364/AO.37.006329.
}
\author{
Soham Mukherjee
}
