% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_stan_data.R
\name{make_stan_data_obs}
\alias{make_stan_data_obs}
\title{Make per observation data needed by STAN model}
\usage{
make_stan_data_obs(
  df,
  stan_data_base,
  use_measured_sigma = TRUE,
  sigma_fallback = 5e-05,
  sigma_floor = 1e-08
)
}
\description{
Make per observation data needed by STAN model
}
