% tinyrox says don't edit this manually, but it can't stop you!
\name{pkg_help}
\alias{pkg_help}
\title{Get help for a package topic as markdown}
\usage{
pkg_help(topic, package)
}
\arguments{
\item{topic}{Character. The help topic name.}

\item{package}{Character. Package name.}
}
\value{
Character string of markdown help text (invisibly). Also printed.
}
\description{
Extracts help documentation and converts it to clean markdown.
}
