% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_saber.R
\name{plot.saber}
\alias{plot.saber}
\title{Plot function for \code{saber} objects

Visualizes the results of a saber analysis with directed graphs.}
\usage{
\method{plot}{saber}(x, test.niveaus = c(1e-10, 1e-05, 0.01, 0.05),
  lty = c("solid", "longdash", "dotdash",
  "dotted")[1:length(test.niveaus)], radius = 48, print.legend = TRUE,
  ...)
}
\arguments{
\item{x}{[\\code{saber.result}] \cr
Result from \code{saberIt}.}

\item{test.niveaus}{[\code{numeric()}] \cr
Which test niveaus should be visualized? At most, 4 niveaus can be specified.}

\item{lty}{[\code{character()}]\cr
Line types for the test niveaus. Must have the same length as \code{test.niveaus},
allowed values are the linetypes of \code{plot.igraph}.}

\item{radius}{[\code{numeric(1)}]
Radius of nodes. Default is 48, should be increased for long algorithm names.}

\item{print.legend}{[\code{logical(1)}]
Should a legend be included? Default is \code{TRUE}.}

\item{...}{[any] \cr
Additional parameters for \code{plot.igraph}.}
}
\value{
\code{invisible(NULL)}, as a side effect, a plot is generated.
}
\description{
Plot function for \code{saber} objects

Visualizes the results of a saber analysis with directed graphs.
}
