% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saber.R
\docType{package}
\name{saber}
\alias{saber}
\alias{saber-package}
\title{saber: Saber data}
\format{Each data set comes in the form of a data frame (with additional tbl
and tbl_df classes from dplyr). Each data frame has its own format.
Additional documentition will be available soon.}
\source{
\url{http://www.icfes.gov.co/investigacion/acceso-a-bases-de-datos}
}
\usage{
data(SB11_20131)
}
\description{
The saber package provides Saber data. Saber are standardized education tests
run by the Colombian Institute for Educational Evaluation (ICFES). By now the
package includes data of the Saber 11 tests since 2005. The Saber 11 are
tests required by every high school student in Colombia to obtain the high
school diploma and are used by the ICFES and the government to assess the
quality of education and by universities for admissions to undergradute
programs. They are applied twice each year to accomodate to both school
calendars. There is a dataset for each test applied since 2005 containing
test results and socio-economic data for every student that took the test
during the given period. Echa dataset is named by the prefix SB11_ followed
by the year (YYYY) and period (1 or 2). See the usage or the example
below.
}
\details{
Each period correspondes to a half of the year, so that the period 1 test
(i.e. first half of the year) is taken by high schools that finish their
academic year on the first half of the year (so-called B calendar) and the
period 2 test is taken by the schools that finish the academic year in the
second half of the year (A calendar). Most schools are A calendar, hence
period 1 data sets are considerably smaller than period 2 data sets.
}
\examples{
data(SB11_20051)
head(SB11_20051)
}

