% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_forward.R
\name{saber_forward}
\alias{saber_forward}
\title{saber_forward}
\usage{
saber_forward(
  wavelength,
  iop,
  water_type = 2,
  view,
  sun,
  optically_shallow,
  h_w = NULL,
  rrs_bottom = NULL,
  verbose = F
)
}
\description{
Fuction repackaged from \code{SABER_forward_fast.R} in codebase. What's the
difference with \code{saber_forward_parametric_conc_wise.R} ?
}
