% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saberIt.R
\name{saberIt}
\alias{saberIt}
\title{Statistical Analysis of Benchmark Results}
\usage{
saberIt(data, perfName, expParName, algoName, replName, clusterFunction)
}
\arguments{
\item{data}{[\code{data.frame}] \cr
Data.frame containing the data to be analyzed}

\item{perfName}{[\code{character(1)}] \cr
Name of the column of data containing the performance values}

\item{expParName}{[\code{character()}] \cr
Names of the columns of data containing the experiment parameters}

\item{algoName}{[\code{character(1)}] \cr
Name of the column of data containing the used algorithm}

\item{replName}{[\code{character(1)}] \cr
Name of the column of data containing the replication number}

\item{clusterFunction}{[\code{character(1)}] \cr
Name of the cluster function to be used. At the moment, only 
clusterFunctionHclust is supported}
}
\value{
A \code{saber}, object.
}
\description{
TODOTODOTODOTODOTODOTODO
}
